+++ pylstemp/__init__.py
@@ -22,4 +22,4 @@ __all__ = [
     "ORIGINAL_LIBRARY_CREDIT",
 ]
 
-__version__ = "1.9.2"
+__version__ = "1.10.0"